requirements.txt
Flask==2.2.2
firebase-admin==6.5.0
numpy==1.23.5
opencv-python==4.7.0.72
tensorflow==2.17.0
imutils==0.5.4
func-timeout==4.3.5
python-dotenv==1.0.1